Fix DomainTag default
