+++ dds/src/implementation/parameter_list_serde/serde_remote_rtps_pim.rs
@@ -1,8 +1,6 @@
-use crate::implementation::data_representation_builtin_endpoints::parameter_id_values::{
-    DEFAULT_DOMAIN_TAG, DEFAULT_EXPECTS_INLINE_QOS,
-};
+use crate::implementation::data_representation_builtin_endpoints::parameter_id_values::DEFAULT_EXPECTS_INLINE_QOS;
 
-#[derive(Debug, PartialEq, Eq, serde::Serialize, derive_more::From)]
+#[derive(Debug, PartialEq, Eq, serde::Serialize)]
 pub struct ExpectsInlineQosSerialize<'a>(pub &'a bool);
 impl Default for ExpectsInlineQosSerialize<'_> {
     fn default() -> Self {
@@ -14,24 +12,5 @@ pub struct ExpectsInlineQosDeserialize(pub bool);
 impl Default for ExpectsInlineQosDeserialize {
     fn default() -> Self {
         Self(DEFAULT_EXPECTS_INLINE_QOS)
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, serde::Serialize)]
-pub struct DomainTag<'a>(pub &'a str);
-impl<'a> Default for DomainTag<'a> {
-    fn default() -> Self {
-        Self(DEFAULT_DOMAIN_TAG)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, serde::Serialize, derive_more::From)]
-pub struct DomainTagSerialize<'a>(pub &'a DomainTag<'a>);
-
-#[derive(Debug, PartialEq, Eq, serde::Deserialize, derive_more::Into)]
-pub struct DomainTagDeserialize(pub String);
-impl Default for DomainTagDeserialize {
-    fn default() -> Self {
-        Self(DEFAULT_DOMAIN_TAG.to_string())
     }
 }
